Updated  Methods
