+++ SauceExamples/webexamples/junit/InstantSauceJunitTest1.java
@@ -1,23 +1,26 @@
-package testng.web.tests;
+package webexamples.junit;
+
+import org.junit.jupiter.api.*;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.openqa.selenium.By;
+import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.remote.DesiredCapabilities;
 import org.openqa.selenium.remote.RemoteWebDriver;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
-import org.testng.annotations.Test;
-import org.testng.Assert;
-import java.lang.reflect.Method;
 import java.net.MalformedURLException;
 import java.net.URL;
 
-public class InstantSauceTestNGTest1 {
+
+public class InstantSauceJunitTest1 {
     private WebDriver driver;
 
     @Test
-    public void shouldOpenSafari(Method method) throws MalformedURLException {
-
+    @DisplayName("shouldOpenSafari()")
+    public void shouldOpenSafari(TestInfo testInfo) throws MalformedURLException {
         /**
          * In this section, we will configure our SauceLabs credentials in order to run our tests on saucelabs.com
          */
@@ -25,10 +28,9 @@ public class InstantSauceTestNGTest1 {
         String sauceAccessKey = "SAUCE_ACCESS_KEY";
 
         /**
-         * In this section, we will configure our test to run a specific
+         * In this section, we will configure our test to run on some specific
          * browser/os combination in Sauce Labs
          */
-
         DesiredCapabilities capabilities = new DesiredCapabilities();
 
         //set your user name and access key to run tests in Sauce
@@ -47,15 +49,8 @@ public class InstantSauceTestNGTest1 {
         capabilities.setCapability("version", "11.1");
 
         //set your test case name so that it shows up in Sauce Labs
-        capabilities.setCapability("name", method.getName());
+        capabilities.setCapability("name", testInfo.getDisplayName());
 
-        /**
-         * In this section, we will set the WebDriver to a Remote driver to run on sauce, and pass the capabilities
-         * we just set. Then we perform som actions on the page before quitting the driver.
-         */
-
-        //create a new Remote driver that will allow your test to send
-        // commands to the Sauce Labs grid so that Sauce can execute your tests
         driver = new RemoteWebDriver(new URL("http://ondemand.saucelabs.com:80/wd/hub"), capabilities);
 
         //navigate to the url of the Sauce Labs Sample app
@@ -81,13 +76,21 @@ public class InstantSauceTestNGTest1 {
         By inventoryPageLocator = By.id("inventory_container");
         wait.until(ExpectedConditions.visibilityOfElementLocated(inventoryPageLocator));
 
-        /**
-         * In this section, we confirm the test ran correctly, howerver we don't post the results to saucelabs.com
-         */
         //Assert that the inventory page displayed appropriately
-        Assert.assertTrue(driver.findElement(inventoryPageLocator).isDisplayed());
+        Boolean result = driver.findElements(inventoryPageLocator).size() > 0;
+        assertTrue(result);
 
-        //Here we tear down the WebDriver session
+        /**
+         * Here we teardown the driver session and send the results to Sauce Labs
+         */
+        if (result){
+            ((JavascriptExecutor)driver).executeScript("sauce:job-result=passed");
+        }
+        else {
+            ((JavascriptExecutor)driver).executeScript("sauce:job-result=failed");
+        }
         driver.quit();
+
     }
+
 }
