+++ SauceExamples/webexamples/junit/JUNIT5_README.md
@@ -1,0 +1,60 @@
+# Java JUnit Jupiter Examples
+
+This directory contains example **Java** scripts using the **JUnit5 (`jupiter`)** framework. Use these scripts for running automated **Selenium tests** on **Sauce Labs**. You can use these scripts to test:
+* Sauce Labs authentication credentials
+* Your IDE setup
+* General Sauce Labs features
+
+Clone or download this repository onto your local machine, import the project into your IDE to begin automated Selenium testing! 
+
+Below are some guided exercsies to get you started.
+##### For Demonstration Purposes Only
+
+> The code provided in these scripts is provided on an "AS-IS” basis without warranty of any kind, either express or implied, including without limitation any implied warranties of condition, uninterrupted use, merchantability, fitness for a particular purpose, or non-infringement. These scripts are provided for educational and demonstration purposes only, and should not be used in production. Issues regarding these scripts should be submitted through GitHub. These scripts are maintained by the Technical Services and Info Development teams at Sauce Labs.
+
+<br />
+
+
+## Prerequisites
+
+See the parent [README](https://github.com/saucelabs-training/Getting-Started-with-Selenium/blob/master/README.md) for instructions on how to complete the following:
+
+* [Install Git](https://github.com/saucelabs-training/Getting-Started-with-Selenium#install-git)
+* [Install IntelliJ](https://github.com/saucelabs-training/Getting-Started-with-Selenium#install-intellij)
+* [Install JDK](https://github.com/saucelabs-training/Getting-Started-with-Selenium#install-jdk)
+* [Setup Selenium Project](https://github.com/saucelabs-training/Getting-Started-with-Selenium#setup-the-project)
+
+## Module List
+1. Run an Automated Test
+2. Test Your App
+3. Configure Cross-Browser Testing
+4. Best Practices Review
+
+<br />
+
+### Module 1: Run Your First Automated Test
+
+1. Checkout branch `01_locate_elements `.
+2. 
+    
+    <br />
+
+### Moudle 2: Configure Environment Variables and Test Against Your Application
+
+1. Checkout branch `02_advanced_locations `.
+2. 
+
+<br />
+    
+### Module 3: Configure Cross-Browser Testing 
+
+1. Checkout branch `03_test_actions`.
+2. 
+
+   <br />
+
+### Module 4: Best Practices for Testing on Sauce Labs
+1. Checkout branch `04_full_journey`.
+2. 
+    <br />
+
